+++ main.tf
@@ -5,14 +5,9 @@ resource "aws_db_parameter_group" "rds_postgres_pg" {
   tags        = var.additional_tags
 }
 
-module "rds_sg" {
-  source               = "./modules/rds-postgres-sg"
-  spark_cluster_sg_ids = var.spark_cluster_sg_ids
-  tamr_vm_sg_id        = var.tamr_vm_sg_id
-  vpc_id               = var.vpc_id
-  security_group_name  = var.security_group_name
-  additional_cidrs     = var.additional_cidrs
-  additional_tags      = var.additional_tags
+resource "aws_db_subnet_group" "rds_postgres_subnet_group" {
+  name       = var.subnet_group_name
+  subnet_ids = var.rds_subnet_ids
 }
 
 resource "aws_db_instance" "rds_postgres" {
@@ -30,11 +25,12 @@ resource "aws_db_instance" "rds_postgres" {
 
   username = var.username
   password = var.password
+  port     = var.db_port
 
-  db_subnet_group_name   = var.subnet_name
+  db_subnet_group_name   = aws_db_subnet_group.rds_postgres_subnet_group.name
   multi_az               = true
   publicly_accessible    = false
-  vpc_security_group_ids = [module.rds_sg.rds_sg_id]
+  vpc_security_group_ids = var.security_group_ids
   parameter_group_name   = aws_db_parameter_group.rds_postgres_pg.name
 
   maintenance_window      = var.maintenance_window